+++ changelog.md
@@ -3,8 +3,12 @@
 ### 0.1.x
 - 0.1.0
   - Added new `Walls` variants
+    - `Basalt Wall`
     - `Calcite Wall`
+    - `Oak Wall`
     - `Polished Basalt`
     - `Smooth Basalt`
   - Added new `Stairs` variants
     - `Calcite Stairs`
+    - `Polished Basalt Stairs`
+    - `Smooth Basalt Stairs`